refactor vacuum mode update modal to VacuumTools to unbloat stepformtoolbox
